+++ prime-angular-frontend/src/app/modules/enrolment/enrolment.module.ts
@@ -23,6 +23,7 @@ import { AccessAgreementCurrentComponent } from './pages/access-agreement-curren
 import {
   AccessAgreementHistoryEnrolmentComponent
 } from './pages/access-agreement-history-enrolment/access-agreement-history-enrolment.component';
+
 import { CollegeCertificationFormComponent } from './shared/components/college-certification-form/college-certification-form.component';
 import { JobFormComponent } from './shared/components/job-form/job-form.component';
 import { EnrolleePageComponent } from './shared/components/enrollee-page/enrollee-page.component';
