+++ prime-angular-frontend/src/app/modules/site-registration/shared/components/site-collection-notice/site-collection-notice.component.html
